[Frontend] Create dashboard page
