+++ package.json
@@ -1,31 +1,33 @@
 {
-    "private": true,
-    "scripts": {
-        "dev": "npm run development",
-        "development": "mix",
-        "watch": "mix watch",
-        "watch-poll": "mix watch -- --watch-options-poll=1000",
-        "hot": "mix watch --hot",
-        "prod": "npm run production",
-        "production": "mix --production"
-    },
-    "devDependencies": {
-        "@babel/preset-react": "^7.17.12",
-        "@chakra-ui/react": "^2.2.1",
-        "@emotion/react": "^11.9.0",
-        "@emotion/styled": "^11.8.1",
-        "@fontsource/nunito": "^4.5.9",
-        "@types/node": "^16.11.39",
-        "@types/react": "^18.0.12",
-        "@types/react-dom": "^18.0.5",
-        "browser-sync": "^2.27.10",
-        "browser-sync-webpack-plugin": "^2.3.0",
-        "framer-motion": "^6.3.11",
-        "laravel-mix": "^6.0.6",
-        "react": "^18.1.0",
-        "react-dom": "^18.1.0",
-        "react-router-dom": "^6.3.0",
-        "ts-loader": "^9.3.0",
-        "typescript": "^4.7.3"
-    }
+  "private": true,
+  "scripts": {
+    "dev": "npm run development",
+    "development": "mix",
+    "watch": "mix watch",
+    "watch-poll": "mix watch -- --watch-options-poll=1000",
+    "hot": "mix watch --hot",
+    "prod": "npm run production",
+    "production": "mix --production"
+  },
+  "devDependencies": {
+    "@babel/preset-react": "^7.17.12",
+    "@chakra-ui/icons": "^2.0.2",
+    "@chakra-ui/react": "^2.2.1",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@fontsource/nunito": "^4.5.9",
+    "@types/node": "^16.11.39",
+    "@types/react": "^18.0.12",
+    "@types/react-dom": "^18.0.5",
+    "browser-sync": "^2.27.10",
+    "browser-sync-webpack-plugin": "^2.3.0",
+    "framer-motion": "^6.3.11",
+    "laravel-mix": "^6.0.6",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-icons": "^4.4.0",
+    "react-router-dom": "^6.3.0",
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.7.3"
+  }
 }
