--- resources/js/components/View.tsx
@@ -1,13 +1,0 @@
-import { BrowserRouter } from "react-router-dom";
-
-import ViewRoutes from "./ViewRoutes";
-
-function View() {
-  return (
-    <BrowserRouter>
-      <ViewRoutes />
-    </BrowserRouter>
-  );
-}
-
-export default View;
