+++ resources/js/pages/App.tsx
@@ -1,0 +1,7 @@
+import View from "./View";
+
+function App() {
+  return <View />;
+}
+
+export default App;
